@@ -36,21 +36,22 @@ dependencies:
   cupertino_icons: ^1.0.8
   
   # Supabase dependencies
-  supabase_flutter: ^2.3.4
-  flutter_secure_storage: ^9.0.0
+  supabase_flutter: ^2.9.1
+  flutter_secure_storage: ^9.2.4
   
   # State management
-  flutter_riverpod: ^2.4.9
+  flutter_riverpod: ^2.6.1
   
   # UI and utilities
+  shared_preferences: ^2.5.3
   lucide_icons: ^0.257.0
-  shared_preferences: ^2.2.2
   fl_chart: ^0.63.0
   shimmer: ^3.0.0
-  connectivity_plus: ^5.0.2
-  firebase_core: ^2.27.0
-  firebase_messaging: ^14.7.10
-  flutter_local_notifications: ^17.0.0
+  connectivity_plus: ^6.1.4
+  firebase_core: ^3.14.0
+  firebase_messaging: ^15.2.7
+  flutter_local_notifications: ^19.3.0
+  http: ^1.2.1
 
 dev_dependencies:
   flutter_test:
