@@ -1,9 +1,17 @@
 import 'package:supabase_flutter/supabase_flutter.dart';
+import '../../config/app_config.dart';
+import 'custom_auth_service.dart';
 import 'supabase_auth_service.dart';
 
 abstract class AuthService {
   // Factory method to get the concrete implementation
-  static AuthService get instance => SupabaseAuthService();
+  static AuthService get instance {
+    if (appConfig.serviceType == ServiceType.supabase) {
+      return SupabaseAuthService();
+    } else {
+      return CustomAuthService();
+    }
+  }
 
   User? get currentUser;
   Stream<AuthState> get authStateChanges;
@@ -25,4 +33,4 @@ abstract class AuthService {
   Future<void> resetPassword(String email);
 
   Future<void> updatePassword(String newPassword);
-} 
+}
